requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
gspread>=5.12.0
PyDrive2>=1.17.0
